requirements.txt
streamlit
openai
pydantic
fastapi[standard]
typing
uvicorn
google-cloud-storage
contextlib
